Fix bug originating from given samples code after code cleanup

+++ FtcRobotController/src/main/java/org/firstinspires/ftc/robotcontroller/external/samples/ConceptAprilTagOptimizeExposure.java
@@ -198,7 +198,7 @@ public class ConceptAprilTagOptimizeExposure extends LinearOpMode
                 exposureControl.setMode(ExposureControl.Mode.Manual);
                 sleep(50);
             }
-            exposureControl.setExposure(exposureMS, TimeUnit.MILLISECONDS);
+            exposureControl.setExposure((long)exposureMS, TimeUnit.MILLISECONDS);
             sleep(20);
 
             // Set Gain.
